chore: rename things to witch
