--- fitter.py
@@ -1,7 +1,0 @@
-"""
-Just a wrapper so that old scripts still work
-"""
-
-from minkasi_jax import fitter
-
-fitter.main()
