+++ unit_tests/RXJ1347_a10.yaml
@@ -1,7 +1,7 @@
 base: "base_unit.yaml"
 name: "RXJ1347"
 model:
-  unit_conversion: "float(mju.get_da(constants['z'])*mju.y2K_RJ(constants['freq'], constants['Te'])*mju.XMpc/mju.me)"
+  unit_conversion: "float(wu.get_da(constants['z'])*wu.y2K_RJ(constants['freq'], constants['Te'])*wu.XMpc/wu.me)"
   structures:
     a10:
       structure: "a10"
