+++ witch/mapmaking.py
@@ -1,6 +1,7 @@
 """
 Functions that wrap useful minkasi recipes
 """
+
 import os
 from typing import Optional
